185495405-custom-data-importers

+++ drivers/hmis_external_apis/app/models/hmis_external_apis/ac_hmis/importers/loaders/esg_funding_assistance_loader.rb
@@ -7,27 +7,11 @@
 # matriculation to new platform
 # creates CustomService and CustomDataElements
 module HmisExternalApis::AcHmis::Importers::Loaders
-  class EsgFundingAssistanceLoader < BaseLoader
-    def perform(rows:)
-      personal_id_by_enrollment_id = Hmis::Hud::Enrollment
-        .where(data_source: data_source)
-        .pluck(:enrollment_id, :personal_id)
-        .to_h
-      records = rows.map do |row|
-        record = model_class.new(default_attrs)
-        record.service_type = custom_service_type
-        record.CustomServiceID = Hmis::Hud::Base.generate_uuid
-        columns.each do |col|
-          col.assign_value(row: row, record: record)
-        end
-        record.personal_id = personal_id_by_enrollment_id.fetch(record.enrollment_id)
-        record.user_id ||= system_user_id
-        record.date_provided = Date.current # FIXME - this isn't right?
-        record
-      end
-
+  class EsgFundingAssistanceLoader < SingleFileLoader
+    def perform
+      records = build_records
       # destroy existing records and re-import
-      model_class.where(data_source: data_source).destroy_all
+      model_class.where(data_source: data_source).destroy_all if clobber
       model_class.import(
         records,
         validate: false,
@@ -38,6 +22,57 @@ module HmisExternalApis::AcHmis::Importers::Loaders
 
     protected
 
+    def filename
+      'ESGFundingAssistance.csv'
+    end
+
+    def build_records
+      personal_id_by_enrollment_id = Hmis::Hud::Enrollment
+        .where(data_source: data_source)
+        .pluck(:enrollment_id, :personal_id)
+        .to_h
+
+      rows.map do |row|
+        record = model_class.new(default_attrs)
+        record.CustomServiceID = Hmis::Hud::Base.generate_uuid
+
+        record.attributes = {
+          enrollment_id: row_value(row, field: 'ENROLLMENTID'),
+          fa_start_date: parse_date(row_value(row, field: 'PAYMENTSTARTDATE')),
+          fa_end_date: parse_date(row_value(row, field: 'PAYMENTENDDATE')),
+          fa_amount: row_value(row, field: 'AMOUNT'),
+          date_created: parse_date(row_value(row, field: 'DATECREATED')),
+          date_updated: parse_date(row_value(row, field: 'DATEUPDATED')),
+          user_id: row_value(row, field: 'USERID') || system_user_id,
+        }
+
+        record.custom_data_elements = build_cdes(row)
+        record.personal_id = personal_id_by_enrollment_id.fetch(record.enrollment_id)
+        record.service_type = custom_service_type
+        record.date_provided = today # FIXME - this isn't right?
+
+        record
+      end
+    end
+
+    def build_cdes(row)
+      [
+        ['FUNDINGSOURCE', :funding_source],
+        ['PAYMENTTYPE', :payment_type],
+      ].map do |field, definition_key|
+        value = row_value(row, field: field)
+        definition = cde_definition(owner_type: model_class.name, key: definition_key)
+        attrs = {
+          owner_type: model_class.name,
+          data_element_definition_id: definition.id,
+          date_created: parse_date(row_value(row, field: 'DATECREATED')),
+          date_updated: parse_date(row_value(row, field: 'DATEUPDATED')),
+          value_string: value,
+        }.merge(default_attrs)
+        Hmis::Hud::CustomDataElement.new(attrs)
+      end
+    end
+
     def custom_service_type
       @custom_service_type ||= begin
         category = Hmis::Hud::CustomServiceCategory.where(
@@ -47,7 +82,7 @@ module HmisExternalApis::AcHmis::Importers::Loaders
         Hmis::Hud::CustomServiceType.where(
           name: 'ESG Funding Assistance',
           custom_service_category_id: category.id,
-          data_source_id: data_source.id
+          data_source_id: data_source.id,
         ).first_or_create!(user_id: system_user_id)
       end
     end
@@ -55,102 +90,5 @@ module HmisExternalApis::AcHmis::Importers::Loaders
     def model_class
       Hmis::Hud::CustomService
     end
-
-    def columns
-      [
-        attr_col('EnrollmentID'),
-        attr_col('PaymentStartDate', map_to: 'FAStartDate'),
-        attr_col('PaymentEndDate', map_to: 'FAEndDate'),
-        attr_col('Amount', map_to: 'FAAmount'),
-        cde_col('FundingSource', definition: funding_source_cde_def, default_attrs: default_attrs),
-        cde_col('PaymentType', definition: payment_type_cde_def, default_attrs: default_attrs),
-        attr_col('DateCreated'),
-        attr_col('DateUpdated'),
-        attr_col('UserID'),
-      ]
-    end
-
-    def funding_source_cde_def
-      @funding_source_cde_def ||= Hmis::Hud::CustomDataElementDefinition.where(
-        owner_type: model_class.name,
-        field_type: :string,
-        key: :funding_source,
-        label: 'Funding Source',
-        data_source_id: data_source.id
-      ).first_or_create!(user_id: system_user_id)
-    end
-
-    def payment_type_cde_def
-      @payment_type_cde_def ||= Hmis::Hud::CustomDataElementDefinition.where(
-        owner_type: model_class.name,
-        field_type: :string,
-        key: :payment_type,
-        label: 'Payment Type',
-        data_source_id: data_source.id
-      ).first_or_create!(user_id: system_user_id)
-    end
-    class BaseColumn
-      attr_accessor :field
-
-      def row_value(row, field:)
-        row[field]&.strip&.presence
-      end
-    end
-
-
-    # 1:1 mapping of field to record attribute
-    class AttributeColumn < BaseColumn
-      attr_accessor :map_to
-      def initialize(field, map_to: nil)
-        self.field = field
-        self.map_to = map_to
-      end
-
-      # assign col from row into record
-      def assign_value(row:, record:)
-        case map_to
-        when String, Symbol
-          record[map_to] = row_value(row, field: field)
-        when nil
-          record[field] = row_value(row, field: field)
-        else
-          raise
-        end
-      end
-    end
-
-    # field is as related CDE on record, importing both record and related CDE recursively
-    class RelatedCommonDataElementColumn < BaseColumn
-      attr_accessor :definition, :default_attrs
-      def initialize(field, definition:, default_attrs:)
-        self.field = field
-        self.definition = definition
-        self.default_attrs = default_attrs
-      end
-
-      # assign col from row into record
-      def assign_value(row:, record:)
-        value =  row_value(row, field: 'FundingSource')
-        return unless value
-
-        cde_attrs = default_attrs.merge({
-          owner_type: model_class.class_name,
-          value_string: value,
-          data_element_definition_id: definition.id,
-          DateCreated: row_value(row, field: 'DateCreated'),
-          DateUpdated: row_value(row, field: 'DateUpdated'),
-        })
-        record.custom_data_elements.build(cde_attrs)
-      end
-    end
-
-    def attr_col(...)
-      AttributeColumn.new(...)
-    end
-
-    def cde_col(...)
-      RelatedCommonDataElementColumn.new(...)
-    end
-
   end
 end
